+++ lib/arcade/api/primitives.rb
@@ -8,7 +8,7 @@ module Arcade
    # persistent http handle to the database
 
     def http
-      break_on = -> (response) { response.status == 500  }
+#     break_on = -> (response) { response.status == 500  }
       @http ||= HTTPX.plugin(:basic_auth).basic_auth(auth[:username], auth[:password])
                   .plugin(:persistent)
                   .plugin(:circuit_breaker)
@@ -41,8 +41,6 @@ module Arcade
 
     # ------------------------------ post data -------------------------------------------------------- #
     def post_data command,  payload
-#      http = HTTPX.plugin(:basic_auth)
- #                 .basic_auth(auth[:username], auth[:password])
       response = http.post( Config.base_uri + command, json:  payload )
       response.raise_for_status
       JSON.parse( response.body, symbolize_names: true )[:result]
@@ -50,17 +48,17 @@ module Arcade
 
     # ------------------------------ transaction      ------------------------------------------------- #
     #
+    # The payload, optional as a JSON, accepts the following parameters:
+    # isolationLevel:  READ_COMMITTED (default)  or REPEATABLE_READ.  (not implemented)
+    #
     def begin_transaction database
       result  = http.post Config.base_uri + "begin/#{database}"
-      @session_id = result.headers["arcadedb-session-id"]
       # returns the session-id
+      result.headers["arcadedb-session-id"] 
     end
 
     # ------------------------------ post transaction ------------------------------------------------- #
-    def post_transaction command, params, session_id= @session_id
-     # http = HTTPX.plugin(:basic_auth)
-     #             .basic_auth(auth[:username], auth[:password])
-     #             .with( headers: { "arcadedb-session-id"=>session }, debug_level: 1)
+    def post_transaction command, params, session_id:
       http_a = http.with(  headers: { "arcadedb-session-id" => session_id } , debug_level: 1)
       response = http_a.post( Config.base_uri + command, json:  params )
       response.raise_for_status
@@ -69,11 +67,10 @@ module Arcade
     end
 
     # ------------------------------ commit           ------------------------------------------------- #
-    def commit database, session_id = @session_id
+    def commit database, session_id:
       http_a = http.with(  headers: { "arcadedb-session-id" => session_id } , debug_level: 1)
       response = http_a.post( Config.base_uri + "commit/#{database}" )
       response.raise_for_status
-      @session_id =  nil
       response.status  #  returns 204 --> success
                        #          403 --> incalid credentials
                        #          500 --> Transaction  not begun
@@ -81,16 +78,12 @@ module Arcade
     end
 
     # ------------------------------ rollback         ------------------------------------------------- #
-    def rollback database, session_id = @session_id, publish_error=true
-    #  http = HTTPX.plugin(:basic_auth)
-    #              .basic_auth(auth[:username], auth[:password])
-    #              .with( headers: { "arcadedb-session-id"=>session_id }, debug_level: 1)
+    def rollback database, session_id: , log: true
       http_a = http.with(  headers: { "arcadedb-session-id" => session_id } , debug_level: 1)
       response = http_a.post( Config.base_uri + "rollback/#{database}" )
       response.raise_for_status
-      @session_id =  nil
-      logger.error  "A Transaction has been rolled back"  # if publish_error
-      response.status
+      logger.info  "A Transaction has been rolled back"   if log
+      response.status    # returns 500 !
     end
   end
  end