+++ CHANGELOG.md
@@ -15,5 +15,11 @@ All notable changes to this project will be documented in this file.
 ## 0.4.0 - 2023.10.28
 - completely remove pg-stuff
 - substitute typhoreous with  HTTPX
-- include dry::monards to process raw-data
 
+## 0.4.1 - 2023.10.31
+- redesign of transactions. 
+- Type.execute performs non idempotent queries  in a (nested) transaction
+- Type.transmit performs non idempotent queries 
+- Type.query performs idempotent queries 
+- Type.create returns a rid
+- Type.insert returns a type-object